requirements.txt
Flask==3.0.0
playwright==1.40.0
python-pptx==0.6.23
beautifulsoup4==4.12.2
requests==2.31.0
gunicorn==21.2.0
